+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version [0.7.0](https://github.com/MC2IT/Agicap.ps1/compare/v0.6.0...v0.7.0)
-- **Suppliers:** added the `New-LegalAddress` cmdlet.
+- **Suppliers:** added the `New-Contact` and `New-LegalAddress` cmdlets.
 
 ## Version [0.6.0](https://github.com/MC2IT/Agicap.ps1/compare/v0.5.1...v0.6.0)
 - **Chart of accounts:** added the `New-AccountingAccount`, `Remove-AccountingAccount`, `Submit-AccountingAccount`, `New-ThirdParty`, `Remove-ThirdParty` and `Submit-ThirdParty` cmdlets.